01_etl/app/requirements.txt
django==3.2
django-environ==0.8.1
whitenoise==5.3.0
flake8==4.0
isort==5.10.1
gunicorn==20.1.0
psycopg2-binary==2.9.3
django-debug-toolbar==3.2.4
djangorestframework==3.13.1
markdown==3.3.6
django-filter==21.1
drf-yasg==1.20.0